+++ src/api/useApi.ts
@@ -5,6 +5,7 @@ export const useApi = createApi({
   baseQuery: fetchBaseQuery({
     baseUrl: import.meta.env.VITE_BASE_URL,
   }),
+  // 헤더 토큰 연결
   endpoints: (builder) => ({
     signIn: builder.mutation<ReturnLoginType, ActionLoginType>({
       query: (data) => ({
@@ -20,11 +21,11 @@ export const useApi = createApi({
         body: data,
       }),
     }),
-    uploadVoice: builder.mutation<any, any>({
+    uploadVoice: builder.mutation<ReturnUploadVoiceType, ActionUploadVoiceType>({
       query: (data) => ({
-        url: '/api/v1/projects/audio/upload',
-        method: 'GET',
-        body: data,
+        url: `/api/v1/projects/${data.projectID}/audio/upload`,
+        method: 'POST',
+        body: data.formData,
       }),
     }),
     getVoiceModel: builder.query<ReturnVoiceModelType, ActionVoiceModelType>({
@@ -68,6 +69,19 @@ interface ActionLoginType {
   email: string;
   password: string;
 }
+interface ActionUploadVoiceType {
+  formData: FormData;
+  projectID: number;
+}
+interface ReturnUploadVoiceType {
+  state: number;
+  result: string;
+  message: string;
+  data: [];
+  error: [];
+}
+[];
+
 interface ReturnVoiceModelType {
   data: [
     {
