+++ src/components/createAvatar/voice/voiceModel/index.tsx
@@ -1,9 +1,9 @@
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useEffect, useRef, useState, ChangeEvent, DragEvent } from 'react';
 import styled from 'styled-components';
 import { useGetVoiceModelQuery, useUploadVoiceMutation } from '../../../../api/useApi';
-import SearchVoiceModelLayout from '../../../../layout/VoiceModel/SearchVoiceModelLayout';
-import VoiceModelFilterButton from '../../../../layout/VoiceModel/VoiceModelFilterButton';
-import VoiceModelListLayout from '../../../../layout/VoiceModel/VoiceModelListLayout';
+import SearchVoiceModelLayout from '../../../../layout/Voice/SearchVoiceModelLayout';
+import VoiceModelFilterButton from '../../../../layout/Voice/VoiceModelFilterButton';
+import VoiceModelListLayout from '../../../../layout/Voice/VoiceModelListLayout';
 
 function VoiceModel() {
   // 음성 모델  전체 리스트 불러오기
@@ -38,15 +38,41 @@ function VoiceModel() {
   interface Uploadtypes {
     // 아직 뭐가 가야하고 뭐가 오는지 모름
   }
-  const [uploadVoice, { data: up, status, isLoading, isError }] = useUploadVoiceMutation();
-  const [uploadFile, setUploadFile] = useState<Uploadtypes | null>(null);
-  function uploadHandler(event: React.ChangeEvent<HTMLInputElement>) {
-    console.log(event.target.files);
-    // uploadVoice({
-    //   projectId: '',
-    //   file: event.target.files?.item(0)
-    // })
-    // 전송되고 아바타로 화면 전환
+  const [onModal, setOnModal] = useState(false);
+  const [uploadFile, { data: url, isLoading: uploading, isError }] = useUploadVoiceMutation();
+  const [audioFile, setAudioFile] = useState<Array<File>>([]);
+
+  // 이미지 파일 처리 input
+  function onInputFile(event: ChangeEvent<HTMLInputElement>) {
+    event.preventDefault();
+    handleFiles(event.target.files!);
+  }
+  // 이미지 파일 처리 ondrop
+  function onDropFiles(event: DragEvent<HTMLDivElement>) {
+    // console.log({ event }, event.dataTransfer.files);
+    event.preventDefault();
+    handleFiles(event.dataTransfer.files);
+  }
+  function handleFiles(files: FileList) {
+    const file: File = files[0];
+    setAudioFile([file]);
+  }
+  const dragOver = (e: DragEvent<HTMLDivElement>) => {
+    e.preventDefault();
+  };
+  function submitHandler(event: React.FormEvent<HTMLFormElement>) {
+    event.preventDefault();
+    let formData = new FormData();
+    console.log(audioFile);
+    formData.append('audioFile', audioFile[0], '테스트파일');
+    // 나중에 프로젝트아이디 연결해야해
+    const projectID = 22;
+    const actionUpload = {
+      formData,
+      projectID,
+    };
+    uploadFile(actionUpload);
+    console.log(url);
   }
 
   // 음성 모델 필터링
@@ -143,7 +169,7 @@ function VoiceModel() {
   }
 
   function selectModel() {
-    // 사용자가 선택한 모델(endData) api전송
+    // 사용자가 선택한 모델(endData) slice전달 그 state를 음성 세부설정에서 전달받음
     console.log('최종 모델 선택');
   }
 
@@ -151,7 +177,17 @@ function VoiceModel() {
     <Container
       onClick={(event: React.MouseEvent<HTMLDivElement, MouseEvent>) => dropdownHandler(event)}
     >
-      <SearchVoiceModelLayout upload={uploadHandler} />
+      <SearchVoiceModelLayout
+        setOnModal={setOnModal}
+        onModal={onModal}
+        audioFile={audioFile}
+        setAudioFile={setAudioFile}
+        onDropFiles={onDropFiles}
+        dragOver={dragOver}
+        onInputFile={onInputFile}
+        submitHandler={submitHandler}
+        uploading={uploading}
+      />
       <VoiceModelFilterButton
         sexButton={sexButton}
         sexFilterHandler={sexFilterHandler}
